+++ src/libtrx/game/stats/common.c
@@ -1,0 +1,219 @@
+#include "debug.h"
+#include "game/game.h"
+#include "game/game_flow.h"
+#include "game/savegame.h"
+#include "game/stats.h"
+#include "log.h"
+#include "utils.h"
+
+bool Stats_IsSecretValid(const int16_t secret_idx)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    if (secret_idx < 0 || secret_idx >= STATS_MAX_SECRETS) {
+        return false;
+    }
+    const uint32_t secret_mask = 1 << secret_idx;
+    return (secret_mask & resume->max_stats.all_secrets_mask) != 0;
+}
+
+bool Stats_HasSecret(const int16_t secret_idx)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    if (secret_idx < 0 || secret_idx >= STATS_MAX_SECRETS) {
+        return false;
+    }
+    const uint32_t secret_mask = 1 << secret_idx;
+    if ((secret_mask & resume->max_stats.all_secrets_mask) == 0) {
+        return false;
+    }
+    return (resume->stats.secret_flags & secret_mask) != 0;
+}
+
+bool Stats_RemoveSecret(const int16_t secret_idx)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    if (secret_idx < 0 || secret_idx >= STATS_MAX_SECRETS) {
+        return false;
+    }
+    const uint32_t secret_mask = 1 << secret_idx;
+    if ((secret_mask & resume->max_stats.all_secrets_mask) == 0) {
+        return false;
+    }
+    if (!(resume->stats.secret_flags & secret_mask)) {
+        return false;
+    }
+    LOG_INFO("Removing secret %d", secret_idx);
+    resume->stats.secret_flags &= ~secret_mask;
+    resume->stats.secret_count--;
+    return true;
+}
+
+bool Stats_AddSecret(const int16_t secret_idx)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    if (secret_idx < 0 || secret_idx >= STATS_MAX_SECRETS) {
+        return false;
+    }
+    const uint32_t secret_mask = 1 << secret_idx;
+    if ((secret_mask & resume->max_stats.all_secrets_mask) == 0) {
+        return false;
+    }
+    if (resume->stats.secret_flags & secret_mask) {
+        return false;
+    }
+    LOG_INFO("Adding secret %d", secret_idx);
+    resume->stats.secret_flags |= secret_mask;
+    resume->stats.secret_count++;
+    return true;
+}
+
+void Stats_UpdateSecrets(LEVEL_STATS *const stats)
+{
+    stats->secret_count = 0;
+    for (int32_t i = 0; i < STATS_MAX_SECRETS; i++) {
+        stats->secret_count += (stats->secret_flags & (1 << i)) ? 1 : 0;
+    }
+}
+
+void Stats_MarkSecretCollected(const ITEM *const item)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    resume->stats.secret_flags |= (uint32_t)(intptr_t)item->data;
+    Stats_UpdateSecrets(&resume->stats);
+}
+
+bool Stats_CheckAllLevelSecretsCollected(void)
+{
+    const RESUME_INFO *const resume =
+        Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    int32_t flags = resume->stats.secret_flags;
+    int32_t count = 0;
+    while (flags != 0) {
+        count += flags & 1;
+        flags >>= 1;
+    }
+
+    return count >= resume->max_stats.max_secret_count;
+}
+
+bool Stats_CheckAllSecretsCollected(GF_LEVEL_TYPE level_type)
+{
+    const FINAL_STATS final_stats = Stats_ComputeFinalStats(level_type);
+    return final_stats.stats.secret_count
+        >= final_stats.max_stats.max_secret_count;
+}
+
+void Stats_AddMedipacksUsed(const double medipack_value)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(Game_GetCurrentLevel());
+    resume->stats.medipacks_used += medipack_value;
+}
+
+void Stats_AddDeath(void)
+{
+    const GF_LEVEL *const current_level = Game_GetCurrentLevel();
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(current_level);
+    resume->stats.death_count++;
+    const int32_t save_slot = Savegame_GetBoundSlot();
+    if (save_slot != -1) {
+        Savegame_UpdateDeathCounters(save_slot, resume->stats.death_count);
+    }
+}
+
+void Stats_UpdateTimer(void)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level != nullptr) {
+        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+        resume->stats.timer++;
+    }
+}
+
+void Stats_AddKill(void)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level != nullptr) {
+        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+        resume->stats.kill_count++;
+    }
+}
+
+void Stats_AddPickup(void)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level != nullptr) {
+        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+        resume->stats.pickup_count++;
+    }
+}
+
+void Stats_AddAmmoHits(void)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level != nullptr) {
+        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+        resume->stats.ammo_hits++;
+    }
+}
+
+void Stats_AddAmmoUsed(void)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level != nullptr) {
+        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+        resume->stats.ammo_used++;
+    }
+}
+
+void Stats_AddDistanceTravelled(const XYZ_32 pos, const XYZ_32 last_pos)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level != nullptr) {
+        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+        resume->stats.distance_travelled += Math_Sqrt(
+            SQUARE(pos.z - last_pos.z) + SQUARE(pos.y - last_pos.y)
+            + SQUARE(pos.x - last_pos.x));
+    }
+}
+
+FINAL_STATS Stats_ComputeFinalStats(const GF_LEVEL_TYPE level_type)
+{
+    FINAL_STATS result = {};
+    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
+    for (int32_t i = 0; i < level_table->count; i++) {
+        const GF_LEVEL *const level = &level_table->levels[i];
+        if (level->type != level_type) {
+            continue;
+        }
+
+        const RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+#define L_ADD(prop) result.prop += resume->prop;
+        L_ADD(stats.kill_count);
+        L_ADD(stats.pickup_count);
+        L_ADD(stats.secret_count);
+        L_ADD(stats.timer);
+        L_ADD(stats.ammo_hits);
+        L_ADD(stats.ammo_used);
+        L_ADD(stats.medipacks_used);
+        L_ADD(stats.distance_travelled);
+        L_ADD(stats.death_count);
+        L_ADD(max_stats.max_kill_count);
+        L_ADD(max_stats.max_pickup_count);
+        L_ADD(max_stats.max_secret_count);
+#undef L_ADD
+    }
+
+    return result;
+}
+
+OBJECT_ID Stats_GetSecretObject(const int32_t secret_idx)
+{
+    const GF_LEVEL *const level = Game_GetCurrentLevel();
+    if (level == nullptr) {
+        return NO_OBJECT;
+    }
+    const LEVEL_MAX_STATS *const stats = Stats_GetLevelMaxStats(level);
+    ASSERT(stats != nullptr);
+    ASSERT(secret_idx >= 0 && secret_idx < STATS_MAX_SECRETS);
+    return stats->secret_objects[secret_idx].assigned_object_id;
+}