+++ src/libtrx/include/libtrx/game/savegame/common.h
@@ -60,9 +60,6 @@ void Savegame_CarryCurrentInfoToNextLevel(
 void Savegame_PersistGameToCurrentInfo(const GF_LEVEL *level);
 void Savegame_ApplyLogicToCurrentInfo(const GF_LEVEL *level);
 
-void Savegame_SetDefaultStats(const GF_LEVEL *level, STATS_COMMON stats);
-STATS_COMMON Savegame_GetDefaultStats(const GF_LEVEL *level);
-
 extern int32_t Savegame_GetSlotCount(void);
 
 #define REGISTER_SAVEGAME_STRATEGY(strategy_)                                  \